// src/types/db.booking.types.ts
export type Database = {
  event_booking: {
    Tables: {
// ...
      bookings: {
        Row: {
          attendant_name: string | null
          booking_date: string | null
          booking_id: string
          booking_status:
            | Database["event_booking"]["Enums"]["booking_status"]
            | null
          created_at: string
          event_status:
            | Database["event_booking"]["Enums"]["event_status"]
            | null
          purchase_date: string
          slot_id: string | null
          ticket_id: string | null
          updated_at: string | null
          user_id: string | null
        }
        Insert: {
          attendant_name?: string | null
          booking_date?: string | null
          booking_id?: string
          booking_status?:
            | Database["event_booking"]["Enums"]["booking_status"]
            | null
          created_at?: string
          event_status?:
            | Database["event_booking"]["Enums"]["event_status"]
            | null
          purchase_date?: string
          slot_id?: string | null
          ticket_id?: string | null
          updated_at?: string | null
          user_id?: string | null
        }
        Update: {
          attendant_name?: string | null
          booking_date?: string | null
          booking_id?: string
          booking_status?:
            | Database["event_booking"]["Enums"]["booking_status"]
            | null
          created_at?: string
          event_status?:
            | Database["event_booking"]["Enums"]["event_status"]
            | null
          purchase_date?: string
          slot_id?: string | null
          ticket_id?: string | null
          updated_at?: string | null
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "bookings_slot_id_fkey"
            columns: ["slot_id"]
            isOneToOne: false
            referencedRelation: "slots"
            referencedColumns: ["slot_id"]
          },
          {
            foreignKeyName: "bookings_ticket_id_fkey"
            columns: ["ticket_id"]
            isOneToOne: false
            referencedRelation: "tickets"
            referencedColumns: ["ticket_id"]
          },
          {
            foreignKeyName: "bookings_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
      event_images: {
        Row: {
          caption: string
          created_at: string
          event_id: string
          image_id: string
          image_url: string
          is_main_image: boolean | null
          metadata: string
          updated_at: string | null
        }
        Insert: {
          caption: string
          created_at?: string
          event_id: string
          image_id?: string
          image_url: string
          is_main_image?: boolean | null
          metadata: string
          updated_at?: string | null
        }
        Update: {
          caption?: string
          created_at?: string
          event_id?: string
          image_id?: string
          image_url?: string
          is_main_image?: boolean | null
          metadata?: string
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "event_images_event_id_fkey"
            columns: ["event_id"]
            isOneToOne: false
            referencedRelation: "events"
            referencedColumns: ["id"]
          }
        ]
      }
// ...
      events: {
        Row: {
          created_at: string
          description: string
          end_date: string
          event_marketing_emailed: boolean | null
          event_status:
            | Database["event_booking"]["Enums"]["event_status"]
            | null
          event_type: Database["event_booking"]["Enums"]["event_type"] | null
          id: string
          location: string
          price: number | null
          recurring_event: boolean | null
          start_date: string
          title: string
          updated_at: string | null
        }
        Insert: {
          created_at?: string
          description: string
          end_date: string
          event_marketing_emailed?: boolean | null
          event_status?:
            | Database["event_booking"]["Enums"]["event_status"]
            | null
          event_type?: Database["event_booking"]["Enums"]["event_type"] | null
          id?: string
          location: string
          price?: number | null
          recurring_event?: boolean | null
          start_date: string
          title: string
          updated_at?: string | null
        }
        Update: {
          created_at?: string
          description?: string
          end_date?: string
          event_marketing_emailed?: boolean | null
          event_status?:
            | Database["event_booking"]["Enums"]["event_status"]
            | null
          event_type?: Database["event_booking"]["Enums"]["event_type"] | null
          id?: string
          location?: string
          price?: number | null
          recurring_event?: boolean | null
          start_date?: string
          title?: string
          updated_at?: string | null
        }
        Relationships: []
      }
// ...
      slots: {
        Row: {
          created_at: string
          end_time: string
          event_id: string
          slot_date: string
          slot_id: string
          start_time: string
          tickets_available: number | null
          updated_at: string | null
        }
        Insert: {
          created_at?: string
          end_time: string
          event_id: string
          slot_date: string
          slot_id?: string
          start_time: string
          tickets_available?: number | null
          updated_at?: string | null
        }
        Update: {
          created_at?: string
          end_time?: string
          event_id?: string
          slot_date?: string
          slot_id?: string
          start_time?: string
          tickets_available?: number | null
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "slots_event_id_fkey"
            columns: ["event_id"]
            isOneToOne: false
            referencedRelation: "events"
            referencedColumns: ["id"]
          }
        ]
      }
      tickets: {
        Row: {
          created_at: string
          event_id: string | null
          slot_id: string | null
          sold_tickets: number | null
          stripe_price_id: string | null
          ticket_id: string
          ticket_name: string
          ticket_price: number | null
          total_tickets: number | null
          updated_at: string | null
        }
        Insert: {
          created_at?: string
          event_id?: string | null
          slot_id?: string | null
          sold_tickets?: number | null
          stripe_price_id?: string | null
          ticket_id?: string
          ticket_name: string
          ticket_price?: number | null
          total_tickets?: number | null
          updated_at?: string | null
        }
        Update: {
          created_at?: string
          event_id?: string | null
          slot_id?: string | null
          sold_tickets?: number | null
          stripe_price_id?: string | null
          ticket_id?: string
          ticket_name?: string
          ticket_price?: number | null
          total_tickets?: number | null
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "tickets_event_id_fkey"
            columns: ["event_id"]
            isOneToOne: false
            referencedRelation: "events"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "tickets_slot_id_fkey"
            columns: ["slot_id"]
            isOneToOne: false
            referencedRelation: "slots"
            referencedColumns: ["slot_id"]
          }
        ]
      }
    }
// ...
  }
}
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof Database["event_booking"]["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof Database["event_booking"]["Enums"]
  ? Database["event_booking"]["Enums"][PublicEnumNameOrOptions]
  : never